Extract twemoji code to separate file
